+++ bash/incl/lcg.bash
@@ -4,7 +4,6 @@ declare -i LCG_SEED
 declare -i LCG_PARAMS_MODULUS
 declare -i LCG_PARAMS_A
 declare -i LCG_PARAMS_C
-declare -i LCG_INTEGER
 
 LCG_SEED=-1
 export LCG_SEED
@@ -77,27 +76,6 @@ lcg() {
     LCG_SEED=${seed_next}
     
     echo "${LCG_SEED}"
-}
-
-lcg_integer() {
-    local -i min=${1:?}
-    local -i max=${2:?}
-    local -i seed res
-    seed=$(lcg)
-    LCG_SEED=${seed}
-    
-    res=$(bc <<< "${seed} % (${max} - ${min} + 1) + ${min}")
-
-    ## Sanity checks
-    if (( res < min || res > max )); then
-        error "INTERNAL: Generated LCG integer is not in [$min,$max]: $res"
-    fi
-
-    ## Export LCG integer as environment variable too
-    # shellcheck disable=SC2034
-    LCG_INTEGER=${res}
-    
-    echo "${res}"
 }
 
 lcg_port() {